#1376: gossiplb: move enum converter out of gossiplb
