+++ src/vt/vrt/collection/balance/gossiplb/gossiplb.cc
@@ -48,6 +48,7 @@
 #include "vt/vrt/collection/balance/gossiplb/gossip_msg.h"
 #include "vt/vrt/collection/balance/gossiplb/gossip_constants.h"
 #include "vt/vrt/collection/balance/gossiplb/criterion.h"
+#include "vt/vrt/collection/balance/lb_args_enum_converter.h"
 #include "vt/context/context.h"
 
 #include <cstdint>
@@ -88,7 +89,7 @@ void GossipLB::inputParams(balance::SpecEntry* spec) {
   bool specified_fanout    = params.find("fanout")    != params.end();
   bool specified_rounds    = params.find("rounds")    != params.end();
 
-  LBArgsEnumConverter<KnowledgeEnum> knowledge_converter_(
+  balance::LBArgsEnumConverter<KnowledgeEnum> knowledge_converter_(
     "knowledge", "KnowledgeEnum", {
       {KnowledgeEnum::UserDefined, "UserDefined"},
       {KnowledgeEnum::Complete,    "Complete"},
@@ -173,7 +174,7 @@ void GossipLB::inputParams(balance::SpecEntry* spec) {
   rollback_      = spec->getOrDefault<bool>("rollback", rollback_);
   target_pole_   = spec->getOrDefault<bool>("targetpole", target_pole_);
 
-  LBArgsEnumConverter<CriterionEnum> criterion_converter_(
+  balance::LBArgsEnumConverter<CriterionEnum> criterion_converter_(
     "criterion", "CriterionEnum", {
       {CriterionEnum::Grapevine,         "Grapevine"},
       {CriterionEnum::ModifiedGrapevine, "ModifiedGrapevine"}
@@ -181,7 +182,7 @@ void GossipLB::inputParams(balance::SpecEntry* spec) {
   );
   criterion_ = criterion_converter_.getFromSpec(spec, criterion_);
 
-  LBArgsEnumConverter<InformTypeEnum> inform_type_converter_(
+  balance::LBArgsEnumConverter<InformTypeEnum> inform_type_converter_(
     "inform", "InformTypeEnum", {
       {InformTypeEnum::SyncInform,  "SyncInform"},
       {InformTypeEnum::AsyncInform, "AsyncInform"}
@@ -189,7 +190,7 @@ void GossipLB::inputParams(balance::SpecEntry* spec) {
   );
   inform_type_ = inform_type_converter_.getFromSpec(spec, inform_type_);
 
-  LBArgsEnumConverter<ObjectOrderEnum> obj_ordering_converter_(
+  balance::LBArgsEnumConverter<ObjectOrderEnum> obj_ordering_converter_(
     "ordering", "ObjectOrderEnum", {
       {ObjectOrderEnum::Arbitrary,        "Arbitrary"},
       {ObjectOrderEnum::ElmID,            "ElmID"},
@@ -200,7 +201,7 @@ void GossipLB::inputParams(balance::SpecEntry* spec) {
   );
   obj_ordering_ = obj_ordering_converter_.getFromSpec(spec, obj_ordering_);
 
-  LBArgsEnumConverter<CMFTypeEnum> cmf_type_converter_(
+  balance::LBArgsEnumConverter<CMFTypeEnum> cmf_type_converter_(
     "cmf", "CMFTypeEnum", {
       {CMFTypeEnum::Original,                   "Original"},
       {CMFTypeEnum::NormByMax,                  "NormByMax"},
